+++ src/components/modal-overlay/modal-overlay.module.css
@@ -1,7 +1,7 @@
 .main {
     width: 100vw;
     height: 200vh;
-    background:#E5E5E5;
+    background: rgba(0, 0, 0, 0.6);
     left: 0px;
     top: 0px;
     position: absolute;
