+++ src/services/order-context.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const OrderContext = React.createContext();
